db_stats/requirements.txt
PyYAML==3.13
psycopg2-binary==2.7.6.1
tqdm==4.28.1
